+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # 0.1 initial version
 
-# 0.2 (current version)
+# 0.2 
 - minor visual tweaks
+
+# 0.3 (current version)
+- saving sorting order state
+- historical line graph on topic progress 
+- retrospective update